requirements.txt
absl_py==1.4.0
attrs==22.2.0
geomloss==0.2.5
gin_config==0.5.0
gym==0.21.0
matplotlib==3.5.3
more_itertools==9.1.0
networkx==2.6.3
numpy==1.21.6
pandas==1.3.5
psutil==5.9.4
recsim==0.2.4
scikit_learn==1.0.2
scipy==1.7.3
simplejson==3.18.1
stable_baselines3==1.5.0
sympy==1.10.1
torch==1.13.1
tqdm==4.64.1
